Use wider layout for Index in 1c build
By the help of \newgeometry{} command provided by the "geometry"
package, provide proper width for the 3-column index.

We have been using "geometry" package for HB papersize since the
support of A4 paper was added.

Now it is also used in 1c builds for all paper sizes.

Enabling "geometry" package requires tweaks of margin sizes in the
preamble.

Margin settings for 2c layout are not touched by this commit.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>