Delegate geometry settings to 'geometry' package
Remove all the magical and direct margin settings which disturb the
geometry package.

For 1c layout, set hcentering=true for ease of reading in PDF readers.
2c layout uses the default margin retio of 2:3 (or 3:2) for two side.

As for vertical margins, set vcentering=true for A4 paper to cope with
its tall paper height.
The other paper sizes use the ratio of 3:4 (top:bottom).

Another option would be to increase textheight for A4 paper.
This will result in different page breaks and might cause unexpected
placement of floating objects, which is not ideal for us.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>