styleguide: Get rid of \VerbatimFootnotes macro
It turns out that this macro is not compatible with the
"footnotebackref" package, which enables hyperlinks from
footnotes to the corresponding footnote symbols in the text.
Furthermore, it is wrong to use the macro in the middle
of a book.
It is supposed to be used next to \begin{document}.

\VerbatimFootnote overrides the definition of \footnote{} based
on the default code of standard LaTeX classes and breaks customization
done by footnotebackref.

In perfbook, this affects footnotes in "Answers to Quick Quizzes"
and Glossary as well as those in the latter part of Style Guide.

As it is not used anywhere else, remove the macro and update the
footnote text accordingly.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>