whymb: Fix description of compiler mischief
In this code snippet, the assertion can not fire due to different
reasons.
When bar() observes b == 1, the smp_mb() assures a == 1.
When bar() observes b == 0, the while loop will loop forever and
the assertion can't be reached.

This was pointed out by Motohiro Kanda in his Japanese translation
of perfbook [1].

[1]: https://sites.google.com/site/kandamotohiro/perfbook-d/perfbookappendixf10

Cc: Motohiro Kanda <kanda.motohiro@gmail.com>
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>