memorder: Reword 'memory-barrier shootdown' -> 'system-wide memory barrier'
As Motohiro pointed out in his translation comment [1], memory barriers
can not be shot down.  Motohiro assumed that it was a typo of
"cache-memory shootdown".

However, in this case, each CPU is only required to issue a memory-
barrier instruction (to wait for ongoing cache activities to complete)
in response to the IPI rather than to flush some cache memory.

This paragraph has since been updated to mention the sys_membarrier
syscall added in Linux kernel v4.3 [2].

Let's borrow the term "system-wide memory barrier" from the commit log.

[1]: https://sites.google.com/view/kandamotohiro2/perfbook-d/perfbookappendixc
[2]: https://git.kernel.org/linus/5b25b13ab08f

Cc: Motohiro Kanda <kanda.motohiro@gmail.com>
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>