Makefile: Cope with GhostScript 9.53.0 or later
Combination of GhostScript 9.53.x and pstricks-base 3.01 requires
an additional option of "gs" to enable transparency.

Current version of a2ping can not cope with this change.

Instead, use the pair of commands "ps2pdf" and "pdfcrop" for .eps
files originated from .tex sources if a new version of gs is detected.

Note:
Due to the use of a deprecated command in pstricks.pro (boilerplate
prologue code), gs 9.53.x emits repetitive warnings of:

   **** WARNING: .setopacityalpha is deprecated (as of 9.53.0) and will be removed in a future release
   **** See .setfillconstantalpha/.setalphaisshape for the improved solution

I'll silence them once I'm sure the stability of new commands in
Makefile.

Reported-by: Дмитрий Дьяченко <dimhen@gmail.com>
Reported-by: Felipe Balbi <balbi@kernel.org>
Tested-by: Felipe Balbi <balbi@kernel.org>
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>