bib: Update missing/broken links in RCU.bib and hw.bib
Add "doi" info to AlexeyGotsman2013ESOPRCU in RCU.bib.

URL of Intelx86MemoryOrdering2007 is now broken.

URLs of imgtec.com in MIPSvII-A-20xx are now broken.
Add MIPSvII-A-2016 with the link to www.mips.com's landing page.
Actual PDF is now on an AWS site.

Annotate a couple of broken URLs with \nolinkurl{}.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>