memorder: Shrink large figures in 2c builds
Figures 15.14 -- 15.16 in 2c builds look unnecessarily large.
Those in 1c builds look nice.
Define \onecolumntextwidth in preamble and use it as the width of
those figures.

There needs a hacky change of float option of Table 15.5 in the next
section as well so that it won't be end up in front of Figure 15.16
in some of the build targets.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>