formal, together: Prevent intrusion of floats into non float snippets
Non-float snippets are used to avoid overloading LaTeX with too
many floating snippets.

In "formal" and "together" chapters, there are a few float objects
which end up in the middle of such code snippets in some of the build
targets.

This commit prevent such unfortunate placement by:

  o shrinking a figure,
  o adding an option to VerbatimN environment, and
  o adding a \FloatBarrier macro

These tweaks are fragile and should be reverted once unintended effect
is observed after the release of Edition 2.

Tested with build targets: -qq, -1cqq, and -1csf.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>