advsync/rt: Adjust placement of Figures 14.12, 14.13, 14.15, and 14.16
For 2c builds, adjust position of Figures 14.12 and 14.13 so that
they (of figure*) are placed on the page where they are mentioned.

Figures 14.15 and 14.16 are close to chapter end and some of build
targets place them (or one of them) in a figure-only page.
Prevent this by using the "H" option in these figures.
This is not enabled on "nq" builds where the figures are _not_
close to chapter end.

Removed a \FloatBarrier macro added in commit c13a368c2d86 ("Add
\FloatBarrier to avoid far-away floats") can be removed now.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>