advsync: Use macro for 'Node A', 'Node B', etc.
This is another workaround of sentences ending with "... Node A.", etc.

If you put the "A" in some macro, such as:

    "... \Node{A}."

and the macro is complex enough, the full stop is implicitly
treated as sentence end.

The definition of \Node{} in perfbook-lt.tex uses math mode for
node names, which can prevent confusion of "A" being an indefinite
article.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>