qqz: Suppress possible orphan heading
As of Edition.2 and current master, heading of Appendix E.7 is
orphaned in perfbook.pdf.

This is because framing of quiz part at the beginning of a section
implies paragraph break and invalids the afterheading-ness of the
first question.

To suppress such an unfortunate result, tweak the \bottomtitlespace
variable of the "titlesec" package for Appendix E.

Positive length of \bottomtitlespace prohibits a section title
to appear in the bottom part of that length and forces the title
to be on the next page.
This should be acceptable in Appendix E.

There exist the same patterns of quick quiz at the beginning of
a section in the main text as well as in -nq builds, which might
cause an orphan heading.

Such cases would be dealt with case by case bases.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>