qqz.sty: Link chapter-wise quizzes from answers in -nq build
Paul's suggestion:

    However, if you click on the either black square at the end of
    the question or the white shadowed square at the end of the answer,
    you get taken back to the small bubble in the text.

    Is it possible that some of our readers will expect that these
    two squares will instead take them back to the question at the
    end of the chapter?

Akira's take on reader's expectation:

    The white boxes at the end of answers are meant as short-cuts to
    go back to the text by a single click for those want to resume
    reading.

So, this commit is a compromise and changes the link of "Quick Quiz"
and black square so that they go back to the chapter-end quiz.
The white box at the end of the answer is kept linking back to the
small bubble in the text.

Also add a tail comment in \QuickQE to avoid an extra skip in front
of the black square.

Suggested-by: Paul E. McKenney <paulmck@kernel.org>
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>