periodcheck: Replace escaped percent symbol first
To prevent "\%" from being treated as a symbol of comment,
replace it with "pct" at the beginning of pattern matching.
(Source .tex files are not modified.)

Also improve regex of label string to cache "chp:foo bar".

Add some comments and do a couple of refactoring by using
(xxx|yyy) and "?" patterns.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>