Auto run periodcheck on 'make'
Run periodcheck on "make" and "make all".
It uses an intermediate target, periodcheck-auto, which depends on
$(targ).
The dependency ensures that periodcheck runs after runlatex completes
in parallel build but doesn't run if runlatex fails.

"make periodcheck" keeps its behavior and does not run pdflatex.

While we are here, improve targets "mslm" and "qq" by adding
dependencies to ensure that the message concerning the obsolete
targets not to be missed.

Suggested-by: Paul E. McKenney <paulmck@kernel.org>
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>