utilities/autodate.sh: Fix potential hung-up on missing tcolorbox
"kpsewhich tcolorbox.sty" can be empty if tcolorbox is not
available.

If this is the case, "grep" in `grep ProvidePackage | sed ...`
will wait forever.

Add existence check of tcolorbox early in the script and error-exit.

To prevent empty autodate.tex from left behind, modify the behavior
of the script so that it doesn't rely on the redirection in Makefile.

Fixes: b2f19231bcc5 ("utilities/autodate.sh: Make newer tcolorbox behave as before")
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>