periodcheck: Check question and exclamation marks as well
It turned out that "?" and "!" have the same behavior as "." with
regard to the spacing behind them.

For example,

    bla bla bla RCU!
    This means bla bla bla.

needs to be coded in the LaTeX source as

    bla bla bla RCU\@!
    This means bla bla bla.

, otherwise space after the "!" would be of normal width.

Also drop the check of annotation of mid-sentence punctuation at
the end of input line.

Our requirement is to break at end-of-sentence punctuation marks.
Mid-sentence punctuation marks at the end of input lines are
allowed as they won't do any harm as long as they are properly
annotated.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>