Add \ignorespaces command to \IfEbookSize and \IfSansSerif
Using \IfEbookSize or \IfSansSerif macro in the middle of a paragraph
can cause extra white spaces around it.
Fix this by adding the \ignorespaces command in their definitions.

Also update the conditional text in styleguide.

As the other conditional macros aren't used in this manner, they are
not modified for the moment.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>