periodcheck: Add colon-related checks
Excerpt of Paul's message:

    > If you have some preference, I can update the scripts to enforce
    > it.

    In this case, I prefer the simple British approach to that of my
    homeland, so please unconditionally capitalize after a colon.

    (Instead of the American approach of doing so only if the word following
    the colon begins a complete sentence, which is not something I want to
    be checking manually, let alone via a script!)

This means colons should always come at the end of LaTeX source lines.
Adding ":" to the existing regex in the check scripts can catch
violations to this convention.

As colons can lead a numbered list of sentences/clauses/phrases,
a colon-specific pattern is also added in cleverefcheck.pl.

Also add a couple of filters to ignore arguments of LaTeX macros
such as \caption{}, \paragraph{}, and \co{}.
Also add a pattern to ignore arguments of \QuickQuizChapter and
\OriginalPublished as these macros can't be broken in-between.

Suggested-by: Paul E. McKenney <paulmck@kernel.org>
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>