datastruct: Reference Figure 10.11 to clarify discussion flow
The preceding paragraph discusses update performance regarding
Figure 10.12.
This paragraph returns to Figure 10.11 and speculates that the rate
of lookup might be affected by the differences of update rates.
To help readers to follow the discussion flow, add a reference to
Figure 10.11 near the beginning of the paragraph.

Also fix a typo in the reworked answer to QQ 10.9.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>