cpu: Mark directories of unused data
This one is a bit tricky because the plots turned out to be unreadable
in the book, nice though they may look on a slide.  The data was
thus hand-applied into tables of numbers.  So this commit leaves two
directories of data as a documentation aid that will hopefully provide
more enlightenment than confusion.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>