cleverefcheck.pl: Add test of listing next to heading
There is a side effect of fancyvrb, or the Verbatim{L|N|U}
environments derived from Verbatim, in that floating "listing"
environments placed just next to chapter/section headings are
treated as page/column-break candidates, and results in a widowed
heading.

Other floating objects such as "figure" and "table" do not
have this issue as long as they are free of fancyvrb.

Add tests to detect "listing" environments who come next to
\chapter/\section/\subsection/... commands.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>