cleverefcheck.pl: Add test of \begin{...} at head of QQA
QQA's heading of "Answer:" looks similar to a section heading,
but it behaves differently when there is some environment next
to it.
Also, if there is an empty line at the head of QQA, "Answer:" can
be widowed.

Add tests to detect such potential break patterns in QQA.

Note: \begin{fcvref} is an exception.  It doesn't induce
page/column break.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>