punctcheck.pl: Fix false positive by stricter pattern for \acr{}
A false positive in defer/rcuintro.tex by punctcheck.pl made
Paul do a wrong band-aide change in commit 8e3364bead00
("defer/rcuintro: Fix ThomasEHart2006a punctuation issue.").

It was because the pattern "\{.+\}" matches all of
"{qsbr}~\cite{ThomasEHart2006a}" and the script saw the
sentence as if it ended with "\IXacrfst{}."

Silence the false positive by using a stricter pattern of
"\{[^\}]+\}".

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>