run{first}latex.sh: Use variable for 'pdflatex' command
In testing a dev version of upstream LaTeX, "pdflatex-dev" is used
in place of "pdflatex".

Using a variable named "LATEX", whose default is pdflatex, in the
scripts helps in such cases.

Update Makefile as well so that the dev version can be specified
as follows:

    make LATEX=pdflatex-dev

Note 1: The trick for setting a default value to a shell variable
using ":" is borrowed from an answer in unix.stackexchange.com:

    https://unix.stackexchange.com/questions/122845/

Note 2: To keep the test of utilities/parallel-latex-regress.sh
working, make copies of the scripts before this change as
utilities/run{first}latex.2022.01.25.sh.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>