memorder: Put \FloatBarrier in front of Section 15.3
Commit e39bbe1b0856 ("memorder: Add example showing store breaking
release-acquire chain") added a couple of code snippets in the
final subsection of Section 15.2.
In 1c builds, this caused the beginning part of Section 15.3 to
appear in the middle of floating code snippets of Section 15.2.

This can be prevented by putting \FloatBarrier in front of Section
15.3, with a (somewhat) wide vertical space left at the end of
Section 15.2.

Note: \FloatBarrier does nothing unless there are some overloaded
float objects. Therefore, 2c and eb builds are not affected.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>