Stick to mfirstuc v2.07
Version 2.08 of LaTeX package "mfirstuc" (indirectly loaded via
"glossaries-extra") is a re-implementation using modern LaTeX3
commands.

Unfortunately, it is not compatible with our indexing macros and
build scripts due to the use of mfirstuc's internal macros.

Newer versions of mfirstuc understand an option [=v2.07] and roll
back to the earlier version [1]. Therefore, as a work around, load
it explicitly with that option before loading glossaries-extra.
It does not affect builds on a system where an earlier version
of mfirstuc is installed.

Link: [1] http://mirrors.ctan.org/macros/latex/contrib/mfirstuc/mfirstuc-manual.html
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>