locking: Adjust coding style of rec_tree_itr.c snippet
perfbook's coding style mostly follows that of the Linux kernel [1].
One exception is the indents of continuation lines, where TABs are
used as many as the leading line, while white spaces are used for
deeper level indents in continuation lines.

This is to keep alignment regardless of TAB:SPC setting.
In two-column layout (2c), code snippets are rendered with
TAB:SPC == 1:2, while in one-column layouts (1c and eb), they are
rendered with TAB:SPC == 1:8.

There is another convention in label names for referencing line
numbers. The base part of such a label often looks like
"ln:<chapter>:<base name of source file>:<some id in the file>"
to make it easy to see where the extracted snippet files (.fcv)
originates.

Stick to these conventions and make other tweaks listed below:

  - Extend the snippet so that the definitions of the structs used in
    the code can be presented.
  - Declare the local variable itr at the beginning of
    tree_for_each_rec().
  - Fold a long line which exceeded the column width of two-column
    layout after the indent adjustment.
  - Mention the source file name where the reference to the snippet
    is made for the first time in the text.

Link: [1] https://www.kernel.org/doc/html/latest/process/coding-style.html
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Cc: Elad Lahav <e2lahav@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>