Codesamples/defer: Make rcuscale.sh use kvm_again.sh
Currently, rcuscale.sh invokes kvm.sh for each run, which results in a
kernel build on each run.  This commit uses kvm.sh with the --build-only
argument to do one build, then uses kvm-again.sh with per-run --bootargs
to use that one kernel for each subsequent run.  This results in a
substantial speedup, with the exact speedup depending on the cost of
kernel builds on the system under test.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>