QAfter: Avoid roundoff and take more measurements
Note that current timespec values have 31 bits of seconds and 30 bits
of nanoseconds, for a total of 61 bits.  This will overflow a double by
some orders of magnitude.  So the key trick is to normalize all of the
measurements relative to the start of the run, which reduces the required
mantissa size from 61 bits to less than 40 bits, assuming runs are only
for a few minutes.  And an IEEE-standard C double easily accommodates
40 bits of mantissa.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>