Stick to LaTeX2e glossaries packages
Commit 55c6bfdbfd44 ("Stick to mfirstuc v2.07") added a rollback
option to the mfirstuc package.

However, release note of glossaries-extra v1.49 [1] mentions related
releases:

    This release is part of a set of inter-connected releases all
    dated 2022-10-14: bib2gls v3.0, glossaries v4.50, mfirstuc v2.08
    and glossaries-extra v1.49. If you update glossaries-extra, make
    sure you update glossaries and mfirstuc as well (and bib2gls, if
    you also use that).

, which suggests that sole rollback of mfirstuc might break
in the future.

To be safe, let's rollback related packages all together.

The glossaries package needs to be loaded in front and needs
options to be consistent with the "acronym" option and the
(slightly different) default behavior of glossaries-extra.

Link: [1] https://www.ctan.org/ctan-ann/id/mailman.1406.1666295591.3715.ctan-ann@ctan.org
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>