Codesamples/depends.mk : Replace non-portable uname '-p' with '-m'
`uname -p` returns 'unknown' for current coreutils.

On GNU coreutils 9.1, the uname --help text says:

uname --help
Usage: uname [OPTION]...
[...]
  -v, --kernel-version     print the kernel version
  -m, --machine            print the machine hardware name
  -p, --processor          print the processor type (non-portable)
  -i, --hardware-platform  print the hardware platform (non-portable)
[...]

Looking deeper, I noticed that '-p' is supposed to be printing the
processor type when available, and not necessarily the architecture.

For printing the architecture, it's recommended to use uname -m, instead.
Even so, it seems like some distros make '-p' to return the same as '-m'
for compatibility purposes.

Anyway, change '-p' to '-m' on depends.mk to avoid getting arch=unknown.

Signed-off-by: Leonardo Bras <leobras.c@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>