gitlab-ci.yml: Add Liberation font families for PDF build
There are a couple of SVG files that use the "Liberation Sans" and
"Liberation Mono" font families, namely defer/multver.svg (Figure 9.15)
and memorder/Alpha.svg (Figure 15.17), among others.
When those font families are not found, Inkscape falls back to
whatever available ones with similar properties.

This dependency is not explicitly mentioned in FAQ-BUILD.txt,
because they are installed as dependencies of required packages
listed for Ubuntu and Fedora.

As for Arch Linux, "ttf-liberation" should cover them.

In Makefile, add check of these font families and advises on missing
nice-to-have font packages.

Expand the answer to #9 in FAQ-BUILD.txt and explain these nice-to-
have fonts for SVG figures.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Cc: Leonardo Bras <leobras.c@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>