utilities/gitlab-ci: Change base machine type to 2-cpu to fix "make check"
Gitlab-CI `make -C CodeSamples/ check` currently fails:

sched_setaffinity: Invalid argument

The problem is that some tests are setting CPU affinity to CPUs 0 and 1,
which means that at least 2 cpus are needed for the test to run properly.
Current (default) saas-linux-small-amd64 instances used for testing on
Gitlab-CI only offer a single CPU, which is not enough.

For the test_codesamples pre-test, change the instance type to
saas-linux-medium-amd64 so the tests can run properly.

Fixes: 9217963c ("utilities/gitlab-ci.yml: Add 'make check' on
test_codesamples")
Signed-off-by: Leonardo Bras <leobras.c@gmail.com>
Reviewed-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>