autodate.sh: Add code for BSD-flavor date command
"date" command on BSD systems has evolved in a way quite different
from that of GNU variants and it is not possible to write a
script compatible with both of them.

Add code for the BSD variant and catch error if both of GNU- and
BSD- options don't work.

For example, the busybox "date" command on alpine doesn't work
with either of them.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>