Use fig2dev as fallback to fig2eps
fig2ps as a distro package is not widely available among Linux distros.
Ubuntu/Debian and Fedora/RHEL/CentOS are outliers in this regard.

Users of other distros need to download a source tar ball from
sourceforge.net (or fossies.org, among others) and manually install
it and its dependencies (read: some of TeX Live packages).

fig2ps would be required if there were LaTeX special text in FIG
files. But that is not the case and fig2dev should suffice.

Add rules in Makefile so that fig2dev can be used for .fig --> .eps
conversion when fig2eps is not found.

Note:
    fig2dev is provided as a package named transfig or fig2dev.
    Alternatively, it might be installed as a requirement of xfig.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>