formal/spinhint: Don't use \qco{} for long message
\qco{} is intended to be used for short snippets and suppresses line
breaks in its argument by using \nbco{} in its definition.
Commit 88fd12717c61 ("formal/spinhint: Use \qco{} instead of``\co{}''")
applied \qco{} for a long error message.
It causes "Underfull \hbox" of large badness preceding it.

In perfbook-1c.log and perfbook-eb.log:

    Underfull \hbox (badness 10000) in paragraph at lines 277--281
    []\T1/ntxtlf/m/n/10 (+20) If you see a mes-sage from [][][][] say-ing:
     []

In perfbook.log:

    Underfull \hbox (badness 2495) in paragraph at lines 277--281
    []\T1/ntxtlf/m/n/10 (+20) If you see a mes-sage from [][][][] say-ing:
     []

Revert the change and permit line breaks in the middle of the quoted
message.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Reviewed-by: SeongJae Park <sj38.park@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>