Use minipage environment in \ebresizeverb
In Ebook-size builds, tens of warnings from varwidth are emitted:

    Package varwidth Warning: Failed to reprocess entire contents
    on input line xxx.

It turns out that this is due to the wrong use of varwidth package.
Instead, the bare minipage environment suffices in \ebresizeverb.

Stop using varwidth and get rid of those warnings.
As a bonus, we no longer need ad-hoc \vspace*{} command at the
bottom of its definition.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>