Silence font warning in sf (sans serif) builds
There are a couple of warnings in sf (sans serif) builds:

    LaTeX Font Warning: Font shape `T1/qhv/m/sl' in size <7.3> not available
    (Font)              Font shape `T1/qhv/m/it' tried instead on input line 367.

This is because the qhv (Helvetica clone) font family doesn't
provide slanted shape to be used in the header area of the page
(chapter and section title) and LaTeX uses its italic shape as a
fallback.

This is OK as italic and slanted shapes would look mostly the same
for sans-serif font families.

Suppress the warning by using the silence package.

Giving "save" option to silence makes it save filtered-out warnings
in $basename.sil.  This is better than discarding silenced warnings.

Add $basename.sil to .gitignore and Makefile's "clean" target.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>