run{first}latex.sh: Reorder checks of fatal error/warning messages
Testing $exitcode and "!pdfTeX error:" or "Emergency stop" early
may ignore an earlier warning that tells the reason of such an
error.

Reorder the checks and make the relevant part of the log file be
printed.

$exitcode should be tested as late as possible.

This improves output when a necessary .sty file is missing:

before:

    or enter new name. (Default extension: sty)

    Enter file name:
    ! Emergency stop.
    <read *>

    l.332 \renewcommand
                       {\sfdefault}{qhv}
    End of file on the terminal!

    Here is how much of TeX's memory you used:
    [...]

after:

    LaTeX Info: Redefining \Bbbk on input line 2810.
    LaTeX Info: Redefining \not on input line 2959.
    )

    ! LaTeX Error: File `biolinum.sty' not found.

    Type X to quit or <RETURN> to proceed,
    or enter new name. (Default extension: sty)

    Enter file name:
    ! Emergency stop.
    [...]

Subsequent runs of latex can also abort due to, e.g., a typo in .bib.
Keep tests of "Emergency stop." and $exitcode in runlatex.sh to
catch such errors.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>