qqz.sty: Work around orphaned section heading
In ebook-size builds, tcolorbox occasionally complains:

    Package tcolorbox Warning: Using nobreak failed. Try to enlarge
    `lines before break' or set page breaks manually on input line
    xxxx.

This warning is accompanied by an unfortunate page/column break just
below a section heading in Appendix E.

Make such ugly breaks less likely by enclosing the \section command
with a pair of pagebreak/nopagebreak hints.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
--
You need to be very lucky to observe this issue.  At the moment,
you can observe it in perfbook-ebsf.pdf built on Ubuntu 22.04.
Section E.18's heading goes to the bottom of page 1277.

This change makes it go to the top of page 1278.
--
 qqz.sty | 3 ++-
 1 file changed, 2 insertions(+), 1 deletion(-)

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>