Use margins/indents for twocolumn as defined in book class
In perfbook.pdf, patches shown in Section 12.1.6.8 exceed the boxes
surrounding them.  Bisection points at commit 1063ddafb476 ("Get rid of
'twocolumn' option in \documentclass").

It turns out that there are a couple of indent and margin settings
the geometry package does not take care of.

Here is a list of such length parameters:

    - \parindent
    - \leftmargini
    - \leftmarginv
    - \leftmarginvi

Those length are set shorter for the twocolumn option in book.cls
and bk10.clo.

Restore the previous behavior by explicitly setting them for twocolumn
builds.

Fixes: 1063ddafb476 ("Get rid of 'twocolumn' option in \documentclass")
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>