Ignore error of unstable Inkscape
Recently, conversions of SVG --> PDF are hit by unstability of
command-line runs of Inkscape under the GNOME session/window manager.

There is an issue ticket at the upstream Inkscape repository [1].

It was closed after a bugfix in glib2, but Fedora 38 has not
received the fix.  Furthermore, pre-release Fedora 39, as well as
openSUSE tumbleweed, still has the very similar symptoms.

As a matter of fact, the conversion is properly finished most
of the time when the Inkscape error happened.

So let's add a make variable IGNORE_INKSCAPE_ERROR and ignore error
code from Inkscape when it is enabled.  It is enabled by default.

[1] Issue #4177 "glib2 2.76.0 breaks Command Line export" at
    https://gitlab.com/inkscape/inkscape/-/issues/

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>