Ignore emergency-save SVG files from Inkscape
When Inkscape is terminated by an error, it saves the SVG file
it was processing into an emergency-save SVG file.

In perfbook, this results in extra SVG files such as:

    - advsync/rt-regimes.svgi.2023_MM_DD_hh_mm_ss.0.svg
    - datastruct/hashzu-a.svgi.2023_MM_DD_hh_mm_ss.0.svg

Ignore those files in successive runs of make and remove them
in "make clean", "make cleanfigs", and "make cleanfigs-svg".

Update .gitignore as well.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>