Makefile: Add rules to use rsvg-convert for SVG --> PDF conversion
Recently, Inkscape has had a number of stability issues in both of
its GUI- and CLI- modes.  We have experienced crashes of CLI-mode
Inkscape caused by a change in glib, which can been resolved
by an update in the GTK3 library now released in GTK 3.24.39.
Commit f307e19f2be0 ("Prevent inkscape from interacting with desktop
manager") was an ugly hack for mitigating such crashes.

There is an alternative converter called rsvg-convert.
It is a light-weight CLI-only converter.

Up until Ubuntu 20.04 LTS, rsvg-convert's coverage of SVG features
was not good enough for some of the figures in perfbook drawn by
using Inkscape.

rsvg-convert has improved since, and version 2.52.5, which is the
version on Ubuntu 22.04 LTS, and later versions do fairly good jobs
in this respect.

So let's use rsvg-convert when it is available.

For Ubuntu and Debian, rsvg-convert belongs to the librsvg2-bin
package.  For Fedora, it is provided in the librsvg2-tools package.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>