count: Update QQ 5.46
Commit cee469f9a7aa ("api: Switch x86 atomics to gcc C11-like
atomics") changed the definition of atomic_set() as follows:

    #define atomic_set(v, i) \
	__atomic_store_n(&(v)->counter, (i), __ATOMIC_RELAXED)

Previously, it was defined for x86 as follows:

    #define atomic_set(v, i)	(((v)->counter) = (i))

QQ 5.46 assumed the old definition and it can confuse those who have
followed the definition under CodeSamples/.

Reword QQ 5.46 to prevent such confusions.

Add a related QQ asking the need of atomic operation for atomic_set().

Reported-by: Hao Lee <haolee.swjtu@gmail.com>
Link: https://lore.kernel.org/perfbook/20230517120110.GA23586@haolee.io/
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>