Plot RCU-test-ratio on-the-fly
RCU-test-ratio.eps has been updated regularly in response to new
releases of mainline Linux kernel.

If such updates were to happen rarely, having the plot as EPS in the
Git repo would make sense.  As for this plot, it makes more sense to
have only the data in the repo and to do the plot on-the-fly.

There is other things to consider.

Using "set term postscript" would require groff as well as gnuplot
for plotting.  By using "set term svg", groff can be omitted from
the requirement list.

Another is the bar style.  As the number of bars is getting larger,
bars with surrounding border can result in moire pattern when viewed
on display.  Bars of "noborder" style should reduce such distracting
effects.

Summary of changes in this commit:

  * Under CodeSamples/formal/data/
    - Remove RCU-test-ratio.eps
    - plot.sh
      - Use "set term svg"
      - Use bar style of "noborder", increase color saturation
      - Other tweaks
    - .gitignore
      - Add *.svg

  * Makefile
    - Add specific recipes for RCU-test-ratio.svg

  * FAQ-BUILD.txt
    - Add gnuplot in the lists of requirements

  * utilities/gitlab-ci.yml
    - Add gnuplot in the list of packages for generate_pdfs

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Cc: Leonardo Bras <leobras.c@gmail.com>  # for gitlab-ci.yml change
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>